Order Service implementation

+++ src/main/java/com/example/HotelSPP/entity/Order.java
@@ -8,13 +8,12 @@ import lombok.ToString;
 import javax.validation.constraints.NotNull;
 import java.util.Date;
 
-
 @Getter
 @Setter
 @ToString
 @Builder
 public class Order {
-    private int id;
+    private long id;
     private @NotNull Date date_created;
     private @NotNull long guest_Id;
     private @NotNull long state_Id;
